+++ samples/counter/src/commonMain/kotlin/com/slack/circuit/sample/counter/CounterPresenter.kt
@@ -6,7 +6,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import com.slack.circuit.internal.runtime.Parcelize
 import com.slack.circuit.retained.rememberRetained
 import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiEvent
@@ -43,7 +42,6 @@ private fun isPrime(value: Int): Boolean {
   return (2..value / 2).none { value % it == 0 }
 }
 
-@Parcelize
 @Serializable
 data object CounterScreen : Screen {
   data class State(val count: Int, val eventSink: (Event) -> Unit = {}) : CircuitUiState
@@ -59,7 +57,6 @@ data object CounterScreen : Screen {
   }
 }
 
-@Parcelize
 @Serializable
 data class PrimeScreen(val number: Int) : Screen {
   data class State(val number: Int, val isPrime: Boolean, val eventSink: (Event) -> Unit = {}) :